Magazine/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
asyncpg
sqlalchemy
pydantic
httpx
python-multipart
bcrypt==4.0.1  # Додайте це вручну
